requirements.txt
langfuse
python-dotenv
transformers
datasets
unsloth
wandb 